Internal change
PiperOrigin-RevId: 305897677
